Fixing secondary sortas attribute conversion
This commit fixes a bug in the script where the "sortas" attribute
in <secondary> elements was not being copied over when Docbook was
converted to HTMLbook. This was caused by a typo where the if clause
was looking for "seondary" instead of "secondary", which meant that
a correctly marked up element would never satisfy the test. The fix
is to spell "secondary" correctly.